edges,nodes,distance functions
